clean way of providing IP-based Virtual Hosting: the mg_get_option() APIs (with mg_get_conn_option() as a new sibling there) will always invoke the option_get user callback (if available) for all ctx and conn-based config parameters, so user code can always decide to deliver alternate values.

+++ test/unit_test.c
@@ -224,7 +224,7 @@ static void test_header_processing()
 	strcpy(buf, input);
 
 	rv = get_request_len(buf, strlen(buf));
-	assert(rv > 0 && rv < strlen(buf));
+	assert(rv > 0 && rv < (int)strlen(buf));
 	assert(strstr(buf + rv, "<HTML><HEAD>") == buf + rv);
 	buf[rv] = 0;
 	p = buf;
